MAJOR REFACTOR - MUCH CLEANER
