--- OSMSpatialite/README.md
@@ -1,7 +1,0 @@
-# OpenMapKit OSMSpatialite
-
-OSMSpatialite is at it's core a C++11 OpenStreetMap data model that
-persists data in a SQLite database with Spatialite GIS extensions.
-
-This library can both run directly in a C++11 application as well as
-in NodeJS via `node-gyp` JavaScript bindings.
