+++ public/README.md
@@ -1,11 +1,11 @@
 # Public Directory
 
 This is the de-facto public directory in which OpenMapKit Server serves
-static files. You can modify the public directory in `settings.js`.
+data and static assets. You can modify the public directory in `settings.js`.
 
 Within the forms directory are several example ODK XForms. These forms
-are available from ODK Collect if you enter the `<server_url>/fs` as
+are available from ODK Collect if you enter the `<server_url>/odk` as
 your server URL.
 
-Submissions to the open `fs` endpoint are written to the `submissions`
+Submissions to the open `odk` endpoint are written to the `submissions`
 directory.
