+++ server.js
@@ -7,30 +7,25 @@ try {
 
 const express = require('express');
 const directory = require('serve-index');
-const fs = require('./ODK/routes/fs');
-const deployments = require('./Deployments/routes');
-const error = require('./ODK/controllers/error-handler');
-const package = require('./package');
+const odk = require('./odk/odk-routes');
+const deployments = require('./deployments/deployment-routes');
+const error = require('./odk/controllers/error-handler');
+const pkg = require('./package');
 const app = express();
 
-app.get('/', function (req, res) {
-    res.status(200).json({
-        name: settings.name,
-        description: settings.description,
-        status: 200,
-        service: package.name,
-        version: package.version
-    });
-});
-
-app.use('/public', express.static(settings.publicDir));
-app.use('/public', directory(settings.publicDir));
-
-// Public File System Endpoint
-app.use('/fs', fs);
+// Basic Info
+app.get('/', info);
+app.get('/info', info);
 
 // Deployments
 app.use('/deployments', deployments);
+
+// Open Data Kit
+app.use('/odk', odk);
+
+// Public Data & Static Assets
+app.use('/public', express.static(settings.publicDir));
+app.use('/public', directory(settings.publicDir));
 
 // Handle errors
 app.use(error);
@@ -38,3 +33,14 @@ app.use(error);
 const port = process.env.PORT || settings.port;
 app.listen(port);
 console.log('OpenMapKit Server is listening on port %s.', port);
+
+
+function info(req, res) {
+    res.status(200).json({
+        name: settings.name,
+        description: settings.description,
+        status: 200,
+        service: pkg.name,
+        version: pkg.version
+    });
+}
